Update for changes in ZIP 234.
In particular, according to ZIP 234 the deployment occurs only once the
NSM subsidy would drop below the legacy subsidy, and so the four-year
periods used for comparison no longer line up with halving periods.

Occurrences of ZSF have also been changed to NSM. The README has been
updated with the new output, and a disclaimer added.

Signed-off-by: Daira-Emma Hopwood <daira@jacaranda.org>